[Feature] Restrict the format of variable code in the variable management module #1851 #1852
