+++ streampark-console/streampark-console-service/src/main/resources/mapper/core/VariableMapper.xml
@@ -18,8 +18,8 @@
 
 -->
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.streampark.console.system.mapper.VariableMapper">
-    <resultMap id="VariableMap" type="org.apache.streampark.console.system.entity.Variable">
+<mapper namespace="org.apache.streampark.console.core.mapper.VariableMapper">
+    <resultMap id="VariableMap" type="org.apache.streampark.console.core.entity.Variable">
         <result column="id" jdbcType="BIGINT" property="id"/>
         <result column="variable_code" jdbcType="VARCHAR" property="variableCode"/>
         <result column="variable_value" jdbcType="VARCHAR" property="variableValue"/>
